refactor: move emoji modules
