+++ apps/meteor/app/reactions/client/methods/setReaction.ts
@@ -1,9 +1,9 @@
 import type { IMessage } from '@rocket.chat/core-typings';
 
+import { emoji } from '../../../../client/lib/emoji';
 import { roomCoordinator } from '../../../../client/lib/rooms/roomCoordinator';
 import { getUser, getUserId } from '../../../../client/lib/user';
 import { Rooms, Subscriptions, Messages } from '../../../../client/stores';
-import { emoji } from '../../../emoji/client';
 
 export const runOptimisticSetReaction = (reaction: string, messageId: IMessage['_id']): void => {
 	if (!getUserId()) {
