+++ apps/meteor/client/contexts/EmojiPickerContext.ts
@@ -1,6 +1,6 @@
 import { createContext, useContext, useMemo } from 'react';
 
-import type { EmojiPickerItem, CategoriesIndexes } from '../../app/emoji/client';
+import type { EmojiPickerItem, CategoriesIndexes } from '../lib/emoji';
 
 type EmojiPickerContextValue = {
 	open: (ref: Element, callback: (emoji: string) => void) => void;
