+++ apps/meteor/client/lib/customEmoji.ts
@@ -1,7 +1,7 @@
 import type { IEmoji } from '@rocket.chat/core-typings';
 import { escapeRegExp } from '@rocket.chat/tools';
 
-import { emoji, removeFromRecent, replaceEmojiInRecent } from '../../app/emoji/client';
+import { emoji, removeFromRecent, replaceEmojiInRecent } from './emoji';
 import { getURL } from '../../app/utils/client';
 
 const isSetNotNull = (fn: () => unknown) => {
