+++ apps/meteor/client/lib/emoji/helpers.ts
