+++ apps/meteor/client/lib/emoji/types.ts
