+++ apps/meteor/client/views/composer/EmojiPicker/EmojiPickerCategoryItem.tsx
@@ -2,7 +2,7 @@ import { IconButton } from '@rocket.chat/fuselage';
 import type { AllHTMLAttributes } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import type { EmojiCategory } from '../../../../app/emoji/client';
+import type { EmojiCategory } from '../../../lib/emoji';
 
 export type EmojiPickerCategoryItemProps = {
 	category: EmojiCategory;
