+++ apps/meteor/client/views/composer/EmojiPicker/SearchingResult.tsx
@@ -7,7 +7,7 @@ import { VirtuosoGrid } from 'react-virtuoso';
 
 import EmojiElement from './EmojiElement';
 import SearchingResultWrapper from './SearchingResultWrapper';
-import type { EmojiItem } from '../../../../app/emoji/client';
+import type { EmojiItem } from '../../../lib/emoji';
 
 /**
  * the `SearchingResults` is missing the previous loadMore function that was implemented before on the latest version of EmojiPicker using the Blaze Template. It can't be implemented because of the issue with react-virtuoso and the custom scrollbars, since its using virtual list its not gonna be an issue rendering bigger results for search
