+++ apps/meteor/client/views/room/contextualBar/Threads/hooks/useNormalizedThreadTitleHtml.ts
@@ -5,9 +5,9 @@ import { useUser, useSetting } from '@rocket.chat/ui-contexts';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { emojiParser } from '../../../../../../app/emoji/client/emojiParser';
 import { filterMarkdown } from '../../../../../../app/markdown/lib/markdown';
 import { MentionsParser } from '../../../../../../app/mentions/lib/MentionsParser';
+import { emojiParser } from '../../../../../lib/emoji/emojiParser';
 
 export const useNormalizedThreadTitleHtml = (mainMessage: IThreadMainMessage) => {
 	const { t } = useTranslation();
