+++ apps/meteor/client/views/room/providers/ComposerPopupProvider.tsx
@@ -9,10 +9,11 @@ import { useMemo, useState } from 'react';
 import type { ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { emoji } from '../../../../app/emoji/client';
 import { slashCommands } from '../../../../app/utils/client';
+import { normalizeUsername } from '../../../../lib/utils/normalizeUsername';
 import { hasAtLeastOnePermission } from '../../../lib/authorization';
 import { pipe } from '../../../lib/cachedStores/pipe';
+import { emoji } from '../../../lib/emoji';
 import { cannedResponsesQueryKeys } from '../../../lib/queryKeys';
 import { Messages, Subscriptions } from '../../../stores';
 import ComposerBoxPopupCannedResponse from '../composer/ComposerBoxPopupCannedResponse';
@@ -27,7 +28,6 @@ import type { ComposerBoxPopupUserProps } from '../composer/ComposerBoxPopupUser
 import type { ComposerPopupContextValue, ComposerPopupOption } from '../contexts/ComposerPopupContext';
 import { ComposerPopupContext, createMessageBoxPopupConfig } from '../contexts/ComposerPopupContext';
 import useCannedResponsesQuery from './hooks/useCannedResponsesQuery';
-import { normalizeUsername } from '../../../../lib/utils/normalizeUsername';
 
 const getToneRegexes = () => {
 	const exactFinalTone = new RegExp('^tone[1-5](?:-[1-5])?:*$');
