+++ apps/meteor/client/views/root/hooks/useNativeEmoji.ts
@@ -1,9 +1,9 @@
 import { useUserPreference } from '@rocket.chat/ui-contexts';
 import { useLayoutEffect } from 'react';
 
-import { emoji } from '../../../../app/emoji/client';
 import { getEmojiConfig } from '../../../../app/emoji-native/lib/getEmojiConfig';
 import { legacyEmojioneMap } from '../../../../app/emoji-native/lib/legacyEmojioneMap';
+import { emoji } from '../../../lib/emoji';
 
 const config = getEmojiConfig(emoji);
