+++ apps/meteor/tests/unit/app/emoji/helpers.spec.ts
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
 import { describe, it, beforeEach, before } from 'mocha';
 
-import { getEmojisBySearchTerm, updateRecent, removeFromRecent, replaceEmojiInRecent } from '../../../../app/emoji/client/helpers';
-import { emoji } from '../../../../app/emoji/client/lib';
 import { getEmojiConfig } from '../../../../app/emoji-native/lib/getEmojiConfig';
+import { getEmojisBySearchTerm, updateRecent, removeFromRecent, replaceEmojiInRecent } from '../../../../client/lib/emoji/helpers';
+import { emoji } from '../../../../client/lib/emoji/lib';
 
 const registerNativeEmojis = () => {
 	const config = getEmojiConfig(emoji);
